try github action for link-checker
