+++ .github/workflows/link-checker.yaml
@@ -1,0 +1,61 @@
+name: "Arc.codes Link Checker"
+
+on: [ pull_request ]
+
+defaults:
+  run:
+    shell: bash
+
+jobs:
+  build:
+    # Setup
+    runs-on: ${{ matrix.os }}
+    strategy:
+      matrix:
+        node-version: [ 14.x, 16.x ]
+        os: [ ubuntu-latest ]
+
+    # Go
+    steps:
+      - name: Check out repo
+        uses: actions/checkout@v2
+
+      - name: Set up Node.js
+        uses: actions/setup-node@v2
+        with:
+          node-version: ${{ matrix.node-version }}
+
+      - name: Env
+        run: |
+          echo "Event name: ${{ github.event_name }}"
+          echo "Git ref:    ${{ github.ref }}"
+          echo "GH actor:   ${{ github.actor }}"
+          echo "SHA:        ${{ github.sha }}"
+          VER=`node --version`; echo "Node ver:   $VER"
+          VER=`npm --version`; echo "npm ver:    $VER"
+
+      - name: Install
+        run: npm install
+
+      - name: Vendor dist files
+        run: npm run dist
+
+      - name: Hydrate
+        run: npx arc hydrate
+        env:
+          CI: true
+
+      - name: check links
+        run: npm run link-checker
+        env:
+          CI: true
+
+      - name: Notify
+        uses: sarisia/actions-status-discord@v1
+        # Only fire alert once
+        if: github.ref == 'refs/heads/main' && failure()
+        with:
+          webhook: ${{ secrets.DISCORD_WEBHOOK }}
+          title: "link checker"
+          color: 0x222222
+          username: GitHub Actions
